// back-end/src/types/Permissions.ts
/**
 * @enum Permissions
 * @description Enum for permissions
 * @property {string} ADMIN - Admin permission
 * @property {string} TRAINER - Trainer permission
 * @property {string} CLIENT - Client permission
 */
enum Permissions {
  ADMIN = 'admin',
  TRAINER = 'trainer',
  CLIENT = 'client',
}

export default Permissions;